Refactor cachingSecretManager
